+++ test/compat/runtime/_.js
@@ -3,4 +3,4 @@
  * @license MIT
  */
 
-export * as arbitrary from "../_arbitraries.js";
+export * as arbitrary from "../../_arbitraries.js";